+++ indico/modules/events/registration/constants.py
@@ -10,3 +10,7 @@ REGISTRATION_PICTURE_SIZE = 1000
 
 #: Max size of the longest edge for registration pictures thumbnails (px)
 REGISTRATION_PICTURE_THUMBNAIL_SIZE = 120
+
+#: Sentinel UUID used to indicate that the user's profile picture should be used as the
+#: registration picture; the actual File is created at form submission time
+PROFILE_PICTURE_SENTINEL = '00000000-0000-0000-0000-000000000002'